Parser impl for basic DIDDoc w/ context & subject
