--- tests/did_doc.rs
@@ -1,2 +1,0 @@
-use did_common::{did, DIDDocument};
-
